Core

+++ src/main/java/com/booleanuk/api/controller/UserController.java
@@ -1,16 +1,17 @@
-package com.booleanuk.api.cinema.controller;
+package com.booleanuk.api.controller;
 
 import com.booleanuk.api.model.User;
+import com.booleanuk.api.repository.PostRepository;
 import com.booleanuk.api.repository.UserRepository;
-import lombok.Getter;
+
+import com.booleanuk.api.response.ErrorResponse;
+import com.booleanuk.api.response.Response;
+import com.booleanuk.api.response.UserListResponse;
+import com.booleanuk.api.response.UserResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 @RestController
 @RequestMapping("users")
@@ -18,46 +19,81 @@ public class UserController {
 
     @Autowired
     private UserRepository userRepository;
-
+    @Autowired
+    private PostRepository postRepository;
 
     @GetMapping
-    public List<User> getAllUsers() {
-        return this.userRepository.findAll();
+    public ResponseEntity<UserListResponse> getAll() {
+        UserListResponse listResponse = new UserListResponse();
+        listResponse.set(this.userRepository.findAll());
+        return ResponseEntity.ok(listResponse);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable int id) {
+    public ResponseEntity<Response<?>> getUserById(@PathVariable int id) {
         User user = this.userRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found by ID"));
-        return ResponseEntity.ok(user);
+                .orElse(null);
+
+        if (user == null) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("not found");
+            return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+        }
+        UserResponse response = new UserResponse();
+        response.set(user);
+        return ResponseEntity.ok(response);
     }
 
     @PostMapping
-    public ResponseEntity<User> createUser(@RequestBody User user) {
-//        user.setCreatedAt(LocalDateTime.now());
-        return ResponseEntity.ok(userRepository.save(user));
+    public ResponseEntity<Response<?>> create(@RequestBody User User ) {
+        UserResponse response = new UserResponse();
+        try {
+            response.set(this.userRepository.save(User));
+        } catch (Exception e) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("Bad request");
+            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable int id, @RequestBody User user) {
+    public ResponseEntity<Response<?>> updateUser(@PathVariable int id, @RequestBody User user) {
         User userToUpdate = this.userRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found by ID"));
+                .orElse(null);
 
-        userToUpdate.setName(user.getName());
+        if (userToUpdate == null) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("not found");
+            return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+        }
+
         userToUpdate.setEmail(user.getEmail());
-        userToUpdate.setPhone(user.getPhone());
+        userToUpdate.setUsername(user.getUsername());
 
-//        userToUpdate.updateUser();
-
-        return new ResponseEntity<User>(userRepository.save(userToUpdate), HttpStatus.CREATED);
+        try {
+            userToUpdate = this.userRepository.save(userToUpdate);
+        } catch (Exception e) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("Bad request");
+            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+        }
+        UserResponse response = new UserResponse();
+        response.set(user);
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
-    @DeleteMapping ResponseEntity<User> deleteUser(@PathVariable int id) {
-        User userToDelete = this.userRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found by ID"));
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Response<?>> deleteUser(@PathVariable int id) {
+        User userToDelete = this.userRepository.findById(id).orElse(null);
+        if (userToDelete == null) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("not found");
+            return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+        }
         this.userRepository.delete(userToDelete);
-        return ResponseEntity.ok(userToDelete);
+        UserResponse response = new UserResponse();
+        response.set(userToDelete);
+        return ResponseEntity.ok(response);
     }
-
-
 }
